Split game state logic into folder
