+++ chariot-server/src/game/mod.rs
@@ -1,10 +1,9 @@
 use std::collections::HashMap;
 use std::net::TcpListener;
-use std::ops::Add;
 use std::thread::{self};
 use std::time::{Duration, Instant};
 
-use chariot_core::networking::ws::{QuestionBody, WSAudienceBoundMessage, WSServerBoundMessage};
+use chariot_core::networking::ws::WSAudienceBoundMessage;
 use chariot_core::networking::Uuid;
 use chariot_core::networking::{
     ClientBoundPacket, ClientConnection, ServerBoundPacket, WebSocketConnection,
@@ -12,10 +11,14 @@ use chariot_core::networking::{
 use chariot_core::player_inputs::InputEvent;
 use chariot_core::GLOBAL_CONFIG;
 
-use crate::chairs::get_player_start_physics_properties;
 use crate::checkpoints::{FinishLine, MajorCheckpoint, MinorCheckpoint};
 use crate::physics::player_entity::PlayerEntity;
 use crate::physics::trigger_entity::TriggerEntity;
+
+use self::phase::*;
+
+mod phase;
+mod voting;
 
 pub struct GameServer {
     listener: TcpListener,
@@ -25,44 +28,6 @@ pub struct GameServer {
     game_state: ServerGameState,
     map: Option<Map>,
 }
-
-pub enum GamePhase {
-    // During this phase, players are in selection screens choosing their chair
-    // and marking ready; they'll be waiting in UI for all of this phase
-    WaitingForPlayerReady,
-    // During this phase, players can see the track, the other players, and have
-    // a countdown until they'll be able to use controls
-    CountingDownToGameStart,
-    // The idea we had was to have idk 30 seconds at the start of the race
-    // before voting starts; players will be driving around normally
-    PlayingBeforeVoting,
-    // >:)))
-    PlayingWithVoting,
-    // Show standings, perhaps a retry button, any other end-of-race stuff
-    AllPlayersDone,
-}
-
-struct WaitingForPlayerReadyState {
-    players_ready: [bool; 4],
-    new_players_joined: Vec<usize>,
-}
-
-struct CountingDownToGameStartState {
-    countdown_end_time: Instant,
-}
-
-struct PlayingBeforeVotingState {
-    voting_start_time: Instant,
-}
-
-struct PlayingWithVotingState {
-    audience_votes: HashMap<Uuid, i32>,
-    current_question: QuestionBody,
-    is_voting_ongoing: bool,
-    vote_close_time: Instant,
-}
-
-struct AllPlayersDoneState {}
 
 pub struct ServerGameState {
     phase: GamePhase,
@@ -91,36 +56,7 @@ impl GameServer {
             ws_server,
             connections: Vec::new(),
             ws_connections: HashMap::new(),
-            game_state: ServerGameState {
-                phase: GamePhase::WaitingForPlayerReady,
-                players: [0, 1, 2, 3]
-                    .map(|num| get_player_start_physics_properties(&String::from("standard"), num)),
-                waiting_for_player_ready_state: WaitingForPlayerReadyState {
-                    players_ready: [false, false, false, false],
-                    new_players_joined: Vec::new(),
-                },
-                counting_down_to_game_start_state: CountingDownToGameStartState {
-                    countdown_end_time: Instant::now(),
-                },
-                playing_before_voting_state: PlayingBeforeVotingState {
-                    voting_start_time: Instant::now(),
-                },
-                playing_with_voting_state: PlayingWithVotingState {
-                    audience_votes: HashMap::new(),
-                    current_question: (
-                        "q".to_string(),
-                        (
-                            "1".to_string(),
-                            "2".to_string(),
-                            "3".to_string(),
-                            "4".to_string(),
-                        ),
-                    ),
-                    is_voting_ongoing: false,
-                    vote_close_time: Instant::now(),
-                },
-                all_players_done_state: AllPlayersDoneState {},
-            },
+            game_state: get_starting_server_state(),
             map: None,
         }
     }
@@ -182,46 +118,6 @@ impl GameServer {
         }
     }
 
-    // creates a websocket for any audience connections
-    fn acquire_any_audience_connections(&mut self) {
-        self.ws_server
-            .set_nonblocking(true)
-            .expect("non blocking should be ok");
-
-        let mut new_uuids: Vec<Uuid> = Vec::new();
-
-        if let Some(stream_result) = self.ws_server.incoming().next() {
-            if let Ok(stream) = stream_result {
-                let id = Uuid::new_v4();
-                let connection = WebSocketConnection::new(stream);
-                self.ws_connections.insert(id, connection);
-                new_uuids.push(id);
-                println!("acquired an audience connection!");
-            }
-        }
-
-        self.ws_server
-            .set_nonblocking(false)
-            .expect("non blocking should be ok");
-
-        for id in new_uuids {
-            let conn = self.ws_connections.get_mut(&id).unwrap();
-
-            conn.push_outgoing_messge(WSAudienceBoundMessage::Assignment(id));
-
-            if matches!(self.game_state.phase, GamePhase::PlayingWithVoting) {
-                if self.game_state.playing_with_voting_state.is_voting_ongoing {
-                    conn.push_outgoing_messge(WSAudienceBoundMessage::Prompt(
-                        self.game_state
-                            .playing_with_voting_state
-                            .current_question
-                            .clone(),
-                    ));
-                }
-            }
-        }
-    }
-
     // handle every packet in received order
     fn process_incoming_packets(&mut self) {
         for (i, connection) in self.connections.iter_mut().enumerate() {
@@ -244,68 +140,6 @@ impl GameServer {
                 }
             }
         }
-    }
-
-    // handle socket data
-    fn process_ws_packets(&mut self) {
-        for (_id, connection) in self.ws_connections.iter_mut() {
-            while let Some(packet) = connection.pop_incoming() {
-                match packet {
-                    WSServerBoundMessage::Vote(id, vote) => {
-                        println!("{} voted for {}", id, vote);
-                        self.game_state
-                            .playing_with_voting_state
-                            .audience_votes
-                            .insert(id, vote);
-                    }
-                }
-            }
-        }
-    }
-
-    // check to see if we need to tally up votes and do something
-    fn check_audience_voting(&mut self) {
-        let state = &mut self.game_state.playing_with_voting_state;
-        if state.is_voting_ongoing && state.vote_close_time < Instant::now() {
-            // time to tally up votes
-            let winner = state
-                .audience_votes
-                .iter()
-                .max_by(|a, b| a.1.cmp(&b.1))
-                .map(|(_key, vote)| vote)
-                .unwrap_or(&0);
-
-            println!("Option {} won!", winner);
-            state.is_voting_ongoing = false;
-            GameServer::broadcast_ws(
-                &mut self.ws_connections,
-                WSAudienceBoundMessage::Winner(*winner),
-            );
-        }
-    }
-
-    // prompts for a question
-    fn start_audience_voting(
-        &mut self,
-        question: String,
-        option1: String,
-        option2: String,
-        option3: String,
-        option4: String,
-        poll_time: Duration,
-    ) {
-        let state = &mut self.game_state.playing_with_voting_state;
-        state.current_question = (question, (option1, option2, option3, option4));
-
-        GameServer::broadcast_ws(
-            &mut self.ws_connections,
-            WSAudienceBoundMessage::Prompt(state.current_question.clone()),
-        );
-
-        state.audience_votes = HashMap::new(); // clear past votes
-        state.is_voting_ongoing = true;
-        state.vote_close_time = Instant::now().add(poll_time);
-        // check on votes in 30 seconds
     }
 
     // sends a message to all connected web clients
